updated setup.sh

+++ setup.sh
@@ -1,8 +1,23 @@
-# Setup docker
-# only if docker is not installed
+#! /bin/bash
+
+# Replaces the domain name in the docker-compose.yml and replacer/data/data.js files, throw an error if no domain is provided
+if [ -z $1 ] 
+then
+  echo "Invalid argument! Syntax: ./setup.sh <YOUR_DOMAIN>"
+  exit 1
+fi
+
+DOMAIN=$1
+
+sed -i.bak "s/corliansa.xyz/${DOMAIN}/g" docker-compose.yml
+sed -i.bak "s/corliansa.xyz/${DOMAIN}/g" replacer/data/data.js
+
+find . -name "*.bak" -type f -delete
+
+# Setup docker if docker is not installed
 if [ ! -x "$(command -v docker)" ]; then
   sudo apt-get update
-  sudo apt-get install \
+  sudo apt-get -y install \
       ca-certificates \
       curl \
       gnupg \
@@ -13,15 +28,19 @@ if [ ! -x "$(command -v docker)" ]; then
     "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
     $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
   sudo apt-get update
-  sudo apt-get install docker-ce docker-ce-cli containerd.io docker-compose-plugin
+  sudo apt-get -y install docker-ce docker-ce-cli containerd.io docker-compose-plugin
 fi
 
 # Setup adguard
-sudo mkdir -p /etc/systemd/resolved.conf.d/
-mv ./adguard/adguardhome.conf /etc/systemd/resolved.conf.d/adguardhome.conf
-mv /etc/resolv.conf /etc/resolv.conf.backup
-ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
-systemctl reload-or-restart systemd-resolved
+if [ ! -f /etc/systemd/resolved.conf.d/adguardhome.conf ] 
+then
+  sudo mkdir -p /etc/systemd/resolved.conf.d/
+  cp ./adguard/adguardhome.conf /etc/systemd/resolved.conf.d/adguardhome.conf
+  mv /etc/resolv.conf /etc/resolv.conf.backup
+  ln -s /run/systemd/resolve/resolv.conf /etc/resolv.conf
+  systemctl reload-or-restart systemd-resolved
+fi
+
 
 # Setup tailscale
 curl -fsSL https://tailscale.com/install.sh | sh
